Debug: Remove key used returning mapped ArtCards.. from destructured props, proptypes and parent div in ArtCard component itself. Only needed for reference in the <ArtCard/> child component returned by map method in Gallery component.
Was previously causing the following error to be logged in client console:
app-index.js:31 Warning: ArtCard: `key` is not a prop. Trying to access it will result in `undefined` being returned. If you need to access the same value within the child component, you should pass it as a different prop. (https://reactjs.org/link/special-props)